Move validations into json package
