+++ src/test/scala/space/kostya/dif/json/SkemaValidationSpec.scala
@@ -1,8 +1,9 @@
-package space.kostya.dif
+package space.kostya.dif.json
 
 import org.scalatest.TryValues
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
+import space.kostya.dif.json.SkemaValidation
 
 class SkemaValidationSpec extends AnyFlatSpec with Matchers {
   "validator based on skema lib" should "return exception for empty json" in {
